+++ mlflow/langchain/chat_agent_langgraph.py
@@ -1,12 +1,12 @@
-import json
 import importlib.metadata
+import json
 import uuid
 from typing import Annotated, Any, Optional, TypedDict
 
 from packaging.version import Version
 
 try:
-    import langgraph
+    import langgraph  # noqa: F401
     from langchain_core.messages import AnyMessage, BaseMessage, convert_to_messages
     from langchain_core.runnables import RunnableConfig
     from langchain_core.runnables.utils import Input
