+++ tests/langgraph/test_chat_agent_langgraph.py
@@ -1,12 +1,6 @@
-import langchain
-import pytest
-from packaging.version import Version
-
-if Version(langchain.__version__) < Version("0.2.0"):
-    pytest.skip("Tests require langchain version 0.2.0 or higher", allow_module_level=True)
-
 import json
 
+import pytest
 from langchain_core.messages import AIMessage, ToolMessage
 
 import mlflow
